refactor: change name folder components
