+++ src/components/ErrorButton/style.scss
