+++ src/components/Main/style.scss
