+++ src/components/Pagination/style.scss
