+++ src/components/Search/style.scss
