+++ src/components/Spinner/style.scss
